Move nodeinfo tests to separate file
